fix: separate attractions page
